RD-main-2/requirements.txt
streamlit==1.55.0
openpyxl
python-bcb
sidrapy
lxml
requests
beautifulsoup4
